all toast added

+++ src/Pages/ShoppingCart/CartItem.jsx
@@ -1,13 +1,35 @@
+import Swal from "sweetalert2";
+
 /* eslint-disable react/prop-types */
-const CartItem = ({ product }) => {
+const CartItem = ({ product, updateCart, setUpdateCart }) => {
   const { _id, image, title, cartPrice } = product;
+  
+  console.log(updateCart);
   const handleDelete = (id) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/cart/${id}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              Swal.fire("Deleted!", "Your file has been deleted.", "success");
+              const remainingCart = updateCart.filter((remCart) => remCart._id !== id)
+              setUpdateCart(remainingCart);
+            }
+          });
+      }
+    });
     console.log(id);
-    fetch(`http://localhost:5000/cart/${id}`, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => console.log(data));
   };
   return (
     <div>
